refactor(mobile): extract queue provider persistence
